Add files via upload
The function of this web widget application is to close an area that is not wanted to appear on the opened page. For example, let's say that when giving a presentation, the question and answer are placed under each other. Thanks to this application, the answer below the question can be turned off or on if desired.
The program gives an error and my request from you is this: When you enlarge the area you want to curtain, the icon (curtain symbol) also grows. Additionally, the curtain size does not remain constant. Please help me. I think this application should definitely be in the current version of Openbord.